refactor: :recycle: More code organization to reduce button.rb size
